+++ crypto/keyring/record.go
@@ -128,6 +128,13 @@ func extractPrivKeyFromRecord(k *Record) (cryptotypes.PrivKey, error) {
 	return extractPrivKeyFromLocal(rl)
 }
 
+// extractPrivKeyFromLocal extracts the private key from a local record.
+// It checks if the private key is available in the provided Record_Local instance.
+// Parameters:
+// - rl: A pointer to a Record_Local instance from which the private key will be extracted.
+// Returns:
+// - priv: The extracted cryptotypes.PrivKey if successful.
+// - error: An error if the private key is not available or if the casting fails.
 func extractPrivKeyFromLocal(rl *Record_Local) (cryptotypes.PrivKey, error) {
 	if rl.PrivKey == nil {
 		return nil, ErrPrivKeyNotAvailable