+++ runtime/v2/module.go
@@ -229,6 +229,7 @@ func ProvideEnvironment(
 // - header.Service
 // - comet.Service
 // - event.Service
+// - store/v2/root.Builder
 //
 // They are all required.  For most use cases these default services bindings should be sufficient.
 // Power users (or tests) may wish to provide their own services bindings, in which case they must
@@ -241,14 +242,16 @@ func DefaultServiceBindings() depinject.Config {
 				stf.NewKVStoreService(actor),
 			)
 		}
-		headerService header.Service = services.NewGenesisHeaderService(stf.HeaderService{})
-		cometService  comet.Service  = &services.ContextAwareCometInfoService{}
-		eventService                 = stf.NewEventService()
+		cometService  comet.Service = &services.ContextAwareCometInfoService{}
+		headerService               = services.NewGenesisHeaderService(stf.HeaderService{})
+		eventService                = services.NewGenesisEventService(stf.NewEventService())
+		storeBuilder                = root.NewBuilder()
 	)
 	return depinject.Supply(
 		kvServiceFactory,
 		headerService,
 		cometService,
 		eventService,
+		storeBuilder,
 	)
 }