+++ server/v2/server_test.go
@@ -20,7 +20,6 @@ import (
 	"cosmossdk.io/server/v2/appmanager"
 	"cosmossdk.io/server/v2/store"
 	storev2 "cosmossdk.io/store/v2"
-	"cosmossdk.io/store/v2/root"
 )
 
 type mockInterfaceRegistry struct{}
@@ -50,17 +49,9 @@ func (*mockApp[T]) InterfaceRegistry() coreserver.InterfaceRegistry {
 	return &mockInterfaceRegistry{}
 }
 
-var _ root.Builder = &mockStoreBuilder{}
-
-type mockStoreBuilder struct{}
-
-func (m mockStoreBuilder) Build(logger log.Logger, config *root.Config) (storev2.RootStore, error) {
-	return nil, nil
+func (*mockApp[T]) GetStore() storev2.RootStore {
+	return nil
 }
-
-func (m mockStoreBuilder) RegisterKey(string) {}
-
-func (m mockStoreBuilder) Get() storev2.RootStore { return nil }
 
 func TestServer(t *testing.T) {
 	currentDir, err := os.Getwd()
@@ -74,18 +65,21 @@ func TestServer(t *testing.T) {
 	cfg := v.AllSettings()
 
 	logger := log.NewLogger(os.Stdout)
+
+	ctx, err := serverv2.SetServerContext(context.Background(), v, logger)
+	require.NoError(t, err)
+
 	grpcServer := grpc.New[transaction.Tx]()
 	err = grpcServer.Init(&mockApp[transaction.Tx]{}, cfg, logger)
 	require.NoError(t, err)
 
-	storeServer := store.New[transaction.Tx](&mockStoreBuilder{})
+	storeServer := store.New[transaction.Tx]()
 	err = storeServer.Init(&mockApp[transaction.Tx]{}, cfg, logger)
 	require.NoError(t, err)
 
 	mockServer := &mockServer{name: "mock-server-1", ch: make(chan string, 100)}
 
 	server := serverv2.NewServer(
-		logger,
 		serverv2.DefaultServerConfig(),
 		grpcServer,
 		storeServer,
@@ -106,7 +100,7 @@ func TestServer(t *testing.T) {
 	require.Equal(t, v.GetString(grpcServer.Name()+".address"), grpc.DefaultConfig().Address)
 
 	// start empty
-	ctx, cancelFn := context.WithCancel(context.TODO())
+	ctx, cancelFn := context.WithCancel(ctx)
 	go func() {
 		// wait 5sec and cancel context
 		<-time.After(5 * time.Second)