+++ simapp/v2/app_di.go
@@ -66,7 +66,6 @@ func AppConfig() depinject.Config {
 			codec.ProvideAddressCodec,
 			codec.ProvideProtoCodec,
 			codec.ProvideLegacyAmino,
-			runtime.ProvideSingletonScopedStoreBuilder,
 		),
 		depinject.Invoke(
 			std.RegisterInterfaces,
@@ -83,8 +82,8 @@ func NewSimApp[T transaction.Tx](
 	var (
 		app          = &SimApp[T]{}
 		appBuilder   *runtime.AppBuilder[T]
-		storeBuilder root.Builder
 		err          error
+		storeBuilder root.Builder
 
 		// merge the AppConfig and other configuration in one config
 		appConfig = depinject.Configs(
@@ -219,6 +218,16 @@ func (app *SimApp[T]) TxConfig() client.TxConfig {
 	return app.txConfig
 }
 
+// GetStore returns the root store.
 func (app *SimApp[T]) GetStore() store.RootStore {
 	return app.store
+}
+
+// Close overwrites the base Close method to close the stores.
+func (app *SimApp[T]) Close() error {
+	if err := app.store.Close(); err != nil {
+		return err
+	}
+
+	return app.App.Close()
 }