+++ simapp/v2/simdv2/cmd/commands.go
@@ -19,7 +19,6 @@ import (
 	"cosmossdk.io/server/v2/cometbft"
 	serverstore "cosmossdk.io/server/v2/store"
 	"cosmossdk.io/simapp/v2"
-	"cosmossdk.io/store/v2/root"
 	confixcmd "cosmossdk.io/tools/confix/cmd"
 
 	"github.com/cosmos/cosmos-sdk/client"
@@ -44,7 +43,6 @@ func newApp[T transaction.Tx](logger log.Logger, viper *viper.Viper) serverv2.Ap
 func initRootCmd[T transaction.Tx](
 	rootCmd *cobra.Command,
 	txConfig client.TxConfig,
-	storeBuilder root.Builder,
 	moduleManager *runtimev2.MM[T],
 ) {
 	cfg := sdk.GetConfig()
@@ -54,13 +52,8 @@ func initRootCmd[T transaction.Tx](
 		genutilcli.InitCmd(moduleManager),
 		debug.Cmd(),
 		confixcmd.ConfigCommand(),
-		NewTestnetCmd(storeBuilder, moduleManager),
+		NewTestnetCmd(moduleManager),
 	)
-
-	logger, err := serverv2.NewLogger(viper.New(), rootCmd.OutOrStdout())
-	if err != nil {
-		panic(fmt.Sprintf("failed to create logger: %v", err))
-	}
 
 	// add keybase, auxiliary RPC, query, genesis, and tx child commands
 	rootCmd.AddCommand(
@@ -72,19 +65,17 @@ func initRootCmd[T transaction.Tx](
 	)
 
 	// wire server commands
-	if err = serverv2.AddCommands(
+	if err := serverv2.AddCommands(
 		rootCmd,
 		newApp,
-		logger,
 		initServerConfig(),
 		cometbft.New(
 			&genericTxDecoder[T]{txConfig},
-			storeBuilder,
 			initCometOptions[T](),
 			initCometConfig(),
 		),
 		grpc.New[T](),
-		serverstore.New[T](storeBuilder),
+		serverstore.New[T](),
 		telemetry.New[T](),
 	); err != nil {
 		panic(err)